Deployed Processor and updated dialogLoader
