+++ MainFrame/Resources/UI/dialogImporter.ui
@@ -7,7 +7,7 @@
     <x>0</x>
     <y>0</y>
     <width>418</width>
-    <height>392</height>
+    <height>215</height>
    </rect>
   </property>
   <property name="windowTitle">
@@ -27,33 +27,15 @@ background-color: rgb(96, 206, 114);
 border-radius: 5px;
 background-color: #51ad61;
 }
-</string>
+
+#btnProcessTimeCard{
+border-radius: 5px;
+color: white;
+background-color: rgb(255, 110, 112);
+}</string>
   </property>
   <layout class="QGridLayout" name="gridLayout">
-   <item row="0" column="0" colspan="2">
-    <widget class="QFrame" name="descFrame">
-     <property name="frameShape">
-      <enum>QFrame::StyledPanel</enum>
-     </property>
-     <property name="frameShadow">
-      <enum>QFrame::Raised</enum>
-     </property>
-     <widget class="QProgressBar" name="progressBar">
-      <property name="geometry">
-       <rect>
-        <x>10</x>
-        <y>250</y>
-        <width>391</width>
-        <height>23</height>
-       </rect>
-      </property>
-      <property name="value">
-       <number>0</number>
-      </property>
-     </widget>
-    </widget>
-   </item>
-   <item row="1" column="0" colspan="2">
+   <item row="0" column="0">
     <widget class="QPushButton" name="importBTN">
      <property name="minimumSize">
       <size>
@@ -79,6 +61,43 @@ background-color: #51ad61;
      </property>
      <property name="text">
       <string> Import</string>
+     </property>
+     <property name="icon">
+      <iconset>
+       <normaloff>../../MainFrame/Resources/Icons/import.svg</normaloff>../../MainFrame/Resources/Icons/import.svg</iconset>
+     </property>
+     <property name="iconSize">
+      <size>
+       <width>30</width>
+       <height>30</height>
+      </size>
+     </property>
+    </widget>
+   </item>
+   <item row="1" column="0">
+    <widget class="QPushButton" name="btnProcessTimeCard">
+     <property name="minimumSize">
+      <size>
+       <width>50</width>
+       <height>80</height>
+      </size>
+     </property>
+     <property name="font">
+      <font>
+       <family>Poppins</family>
+       <pointsize>18</pointsize>
+       <weight>75</weight>
+       <bold>true</bold>
+      </font>
+     </property>
+     <property name="cursor">
+      <cursorShape>PointingHandCursor</cursorShape>
+     </property>
+     <property name="styleSheet">
+      <string notr="true"/>
+     </property>
+     <property name="text">
+      <string>Process Time Card</string>
      </property>
      <property name="icon">
       <iconset>
