+++ MainFrame/Resources/UI/timecard.ui
@@ -24,7 +24,7 @@
 	border-radius: 5px;
 }
 
-#btnTimeSheet{
+#btnTimeSheet, #importBTN{
 border-radius: 5px;
 background-color: &quot;#30d977&quot;;
 }</string>
@@ -237,7 +237,7 @@ QTableWidget {
      <string notr="true"/>
     </property>
     <property name="text">
-     <string>Create Time Sheet</string>
+     <string>Process Time Sheet</string>
     </property>
     <property name="iconSize">
      <size>
@@ -757,11 +757,6 @@ border-radius: 5px;
        <pointsize>12</pointsize>
       </font>
      </property>
-     <item>
-      <property name="text">
-       <string>data</string>
-      </property>
-     </item>
     </widget>
     <widget class="QLabel" name="lblFromCC">
      <property name="geometry">
@@ -784,6 +779,39 @@ border-radius: 5px;
       <string>From:</string>
      </property>
     </widget>
+   </widget>
+   <widget class="QPushButton" name="importBTN">
+    <property name="geometry">
+     <rect>
+      <x>20</x>
+      <y>270</y>
+      <width>211</width>
+      <height>41</height>
+     </rect>
+    </property>
+    <property name="font">
+     <font>
+      <family>Poppins</family>
+      <pointsize>12</pointsize>
+      <weight>75</weight>
+      <bold>true</bold>
+     </font>
+    </property>
+    <property name="cursor">
+     <cursorShape>PointingHandCursor</cursorShape>
+    </property>
+    <property name="styleSheet">
+     <string notr="true"/>
+    </property>
+    <property name="text">
+     <string>DAT Import</string>
+    </property>
+    <property name="iconSize">
+     <size>
+      <width>22</width>
+      <height>22</height>
+     </size>
+    </property>
    </widget>
   </widget>
   <widget class="QFrame" name="header">
